Endret DialogMessageProcessor til DialogMessageGenerator
