+++ src/main/kotlin/no/nav/helsemelding/messagegenerator/scheduler/SchedulerService.kt
@@ -2,13 +2,13 @@ package no.nav.helsemelding.messagegenerator.scheduler
 
 import kotlinx.coroutines.CoroutineScope
 import no.nav.helsemelding.messagegenerator.config.Config
-import no.nav.helsemelding.messagegenerator.processor.DialogMessageProcessor
-import no.nav.helsemelding.messagegenerator.processor.IncomingMessageProducer
+import no.nav.helsemelding.messagegenerator.generator.DialogMessageGenerator
+import no.nav.helsemelding.messagegenerator.generator.IncomingMessageProducer
 
 class SchedulerService(
     scope: CoroutineScope,
     config: Config,
-    dialogMessageProcessor: DialogMessageProcessor,
+    dialogMessageGenerator: DialogMessageGenerator,
     incomingMessageProducer: IncomingMessageProducer
 ) {
     val dialogMessages = ManagedScheduler(
@@ -17,7 +17,7 @@ class SchedulerService(
         initialInterval = config.kafka.topics.dialogMessage.interval,
         scope = scope,
         action = {
-            dialogMessageProcessor.processMessages(scope)
+            dialogMessageGenerator.generateMessages(scope)
         }
     )
