+++ src/test/kotlin/no/nav/helsemelding/messagegenerator/generator/DialogMessageGeneratorSpec.kt
@@ -1,0 +1,87 @@
+package no.nav.helsemelding.messagegenerator.generator
+
+import io.kotest.core.spec.style.StringSpec
+import io.kotest.inspectors.forAll
+import io.kotest.inspectors.forSingle
+import io.kotest.matchers.collections.shouldBeOneOf
+import io.kotest.matchers.collections.shouldHaveSize
+import io.kotest.matchers.equality.shouldBeEqualUsingFields
+import io.kotest.matchers.string.shouldContain
+import io.kotest.matchers.string.shouldNotContain
+import io.kotest.matchers.types.shouldBeInstanceOf
+import no.nav.helsemelding.messagegenerator.model.InvalidDialogMessage
+import no.nav.helsemelding.messagegenerator.model.ValidDialogMessage
+import no.nav.helsemelding.messagegenerator.util.nowWithOffset
+import no.nav.helsemelding.messagegenerator.util.readFileToString
+import no.nav.helsemelding.messagegenerator.util.replaceInTemplate
+import kotlin.uuid.Uuid
+
+class DialogMessageGeneratorSpec : StringSpec(
+    {
+        "Replaces values in xml template" {
+            val xml = readFileToString("templates/dialogMessage.xml")!!
+
+            xml shouldContain "{genDate}"
+            xml shouldContain "{messageId}"
+            xml shouldContain "{senderHerId}"
+            xml shouldContain "{receiverHerId}"
+            xml shouldContain "{patientName}"
+            xml shouldContain "{message}"
+
+            val uuid = Uuid.random().toString()
+            val name = "Name"
+            val message = "Message"
+
+            val genDate = nowWithOffset()
+            val myMap = mapOf(
+                "{genDate}" to genDate,
+                "{messageId}" to uuid,
+                "{senderHerId}" to FAGSYSTEM_HERID,
+                "{receiverHerId}" to EPJ_HERID,
+                "{patientName}" to name,
+                "{message}" to message
+            )
+
+            val replacedXml = replaceInTemplate(xml, myMap)
+
+            replacedXml shouldNotContain "{genDate}"
+            replacedXml shouldNotContain "{messageId}"
+            replacedXml shouldNotContain "{senderHerId}"
+            replacedXml shouldNotContain "{receiverHerId}"
+            replacedXml shouldNotContain "{patientName}"
+            replacedXml shouldNotContain "{message}"
+
+            replacedXml shouldContain "<GenDate>$genDate</GenDate>"
+            replacedXml shouldContain "<MsgId>$uuid</MsgId>"
+            replacedXml shouldContain "<Id>$FAGSYSTEM_HERID</Id>"
+            replacedXml shouldContain "<Id>$EPJ_HERID</Id>"
+            replacedXml shouldContain "<GivenName>$name</GivenName>"
+            replacedXml shouldContain "<Sporsmal>$message</Sporsmal>"
+        }
+
+        val xml = "<MsgHead></MsgHead>"
+
+        "Number 1 through 9 should create a valid dialog message" {
+            (1..9).toList().forAll { number ->
+                nextDialogMessage(xml, number).forSingle { dialogMessage ->
+                    dialogMessage.shouldBeInstanceOf<ValidDialogMessage>()
+                }
+            }
+        }
+
+        "Number 10 should create an invalid dialog message" {
+            nextDialogMessage(xml, 10).forSingle { invalidDialogMessage ->
+                invalidDialogMessage.shouldBeInstanceOf<InvalidDialogMessage>()
+                invalidDialogMessage.id.shouldBeOneOf(null, "", "1234-abcd")
+            }
+        }
+
+        "Number 11 should create two dialog messages with the same uuid" {
+            val dialogMessages = nextDialogMessage(xml, 11).shouldHaveSize(2)
+
+            val firstMessage = dialogMessages.elementAt(0).shouldBeInstanceOf<ValidDialogMessage>()
+            val secondMessage = dialogMessages.elementAt(1).shouldBeInstanceOf<ValidDialogMessage>()
+            firstMessage shouldBeEqualUsingFields secondMessage
+        }
+    }
+)
